+++ fabric-1.19.4/src/main/java/io/github/kurrycat/mpkmod/compatibility/fabric_1_19_4/FunctionCompatibility.java
@@ -8,6 +8,7 @@ import io.github.kurrycat.mpkmod.util.BoundingBox3D;
 import io.github.kurrycat.mpkmod.util.Debug;
 import io.github.kurrycat.mpkmod.util.Vector2D;
 import io.github.kurrycat.mpkmod.util.Vector3D;
+import io.github.kurrycat.mpkmod.util.ScissorBox;
 import io.github.kurrycat.mpknetapi.common.network.packet.MPKPacket;
 import io.netty.buffer.Unpooled;
 import net.fabricmc.fabric.api.client.networking.v1.ClientPlayNetworking;
@@ -451,21 +452,5 @@ public class FunctionCompatibility implements FunctionHolder,
 
     public void endSection() {
         MinecraftClient.getInstance().getProfiler().pop();
-    }
-
-    private static class ScissorBox {
-        public double x, y, w, h;
-
-        public ScissorBox(double x, double y, double w, double h) {
-            this.x = x;
-            this.y = y;
-            this.w = w;
-            this.h = h;
-        }
-
-        public boolean contains(Vector2D point) {
-            return x <= point.getX() && point.getX() < x + w &&
-                    y <= point.getY() && point.getY() < y + h;
-        }
     }
 }
