+++ forge-1.8.9/src/main/java/io/github/kurrycat/mpkmod/compatibility/forge_1_8/FunctionCompatibility.java
@@ -5,6 +5,7 @@ import io.github.kurrycat.mpkmod.util.BoundingBox3D;
 import io.github.kurrycat.mpkmod.util.Debug;
 import io.github.kurrycat.mpkmod.util.Vector2D;
 import io.github.kurrycat.mpkmod.util.Vector3D;
+import io.github.kurrycat.mpkmod.util.ScissorBox;
 import io.github.kurrycat.mpknetapi.common.network.packet.MPKPacket;
 import net.minecraft.block.Block;
 import net.minecraft.block.properties.IProperty;
@@ -463,21 +464,5 @@ public class FunctionCompatibility implements FunctionHolder,
      */
     public void endSection() {
         Minecraft.getMinecraft().mcProfiler.endSection();
-    }
-
-    private static class ScissorBox {
-        public double x, y, w, h;
-
-        public ScissorBox(double x, double y, double w, double h) {
-            this.x = x;
-            this.y = y;
-            this.w = w;
-            this.h = h;
-        }
-
-        public boolean contains(Vector2D point) {
-            return x <= point.getX() && point.getX() < x + w &&
-                    y <= point.getY() && point.getY() < y + h;
-        }
     }
 }
